fix: Chevron automirrored 图标不可用改 KeyboardArrow + 海报绘制 Int→Float
- CalendarScreen: automirrored.filled.ChevronLeft/Right 在依赖图标集中不存在
  → 改用 Icons.Filled.KeyboardArrowLeft/Right（core 必备图标）
- PosterGenerator: drawDishCell 行列坐标与 footerY 由 Int 显式 toFloat，
  避免 Canvas Float 参数类型不匹配

+++ app/src/main/java/com/family/menu/util/PosterGenerator.kt
@@ -79,8 +79,8 @@ object PosterGenerator {
             shown.forEachIndexed { i, line ->
                 val col = i % 2
                 val row = i / 2
-                val x = 56 + col * (464 + 40)
-                val y = gridStartY + row * cellH
+                val x = (56 + col * (464 + 40)).toFloat()
+                val y = (gridStartY + row * cellH).toFloat()
                 drawDishCell(canvas, line, x, y)
             }
             if (lines.size > 12) {
